An example FunctionalSingleCellExperiment

Contains sample data for a functional single-cell experiment.

fsce_small

Format

An object of class FunctionalSingleCellExperiment of length 2.

Source

Generated by data-raw/fsce_small.R.

Details

There are two SingleCellExperiments:

rnaseq: a SingleCellExperiment object containing a single-cell mRNA sequencing (10x Genomics V2 3 prime) cell/gene counts matrix in the counts slot. Log-normalized counts are in the logcounts slot.

haircut: a SingleCellExperiment object containing a Haircut experiment using DNA repair oligos. Counts are in the counts slot, and centered log-ratio normalized counts are in the logcounts slot.

Haircut repair substrates include:

A "normal" DNA oligo (Normal_)

a ribonucleotide (rG) at position 45 (Ribo_)

an Abasic site at position 45 (Abasic_)

a uracil:adenosine base-pair, with uracil at position 45 (Uracil_)

a uracil:guanosine base-pair, with uracil at position 45 (GU_)

Dimensionality reduction calculations are in the rnaseq experiment, including PCA, UMAP, and TSNE slots. In addition, clusters determined by cluster_kmeans() and cluster_leiden() are available as colData.

Examples

fsce_small

#> A FunctionalSingleCellExperiment object of 2 listed
#>  experiments with user-defined names and respective classes. 
#>  Containing an ExperimentList class object of length 2: 
#>  [1] rnaseq: SingleCellExperiment with 9462 rows and 250 columns 
#>  [2] haircut: SingleCellExperiment with 426 rows and 250 columns 
#> Features: 
#>  experiments() - obtain the ExperimentList instance 
#>  colData() - the primary/phenotype DataFrame 
#>  sampleMap() - the sample availability DataFrame 
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment 
#>  *Format() - convert into a long or wide DataFrame 
#>  assays() - convert ExperimentList to a SimpleList of matrices

# Individual experiments
fsce_small[["rnaseq"]]

#> class: SingleCellExperiment 
#> dim: 9462 250 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames(9462): NOC2L HES4 ... AC004556.1 AC240274.1
#> rowData names(0):
#> colnames(250): TGCGGGTGTAGAGTGC TGGTTCCCATCTATGG ... CTCGAAAAGTTCGATC
#>   TGACAACGTCGCATAT
#> colData names(4): cell_id k_cluster leiden_cluster cell_cycle
#> reducedDimNames(3): PCA UMAP TSNE
#> spikeNames(0):

fsce_small[["haircut"]]

#> class: SingleCellExperiment 
#> dim: 426 250 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames(426): Abasic_1 Abasic_10 ... riboG_8 riboG_9
#> rowData names(2): hairpin position
#> colnames(250): TGCGGGTGTAGAGTGC TGGTTCCCATCTATGG ... CTCGAAAAGTTCGATC
#>   TGACAACGTCGCATAT
#> colData names(1): cell_id
#> reducedDimNames(0):
#> spikeNames(0):

# Gene and activity names (first 5 items)
rownames(fsce_small[["rnaseq"]])[1:5]

#> [1] "NOC2L"   "HES4"    "ISG15"   "TNFRSF4" "SDF4"   

rownames(fsce_small[["haircut"]])[1:5]

#> [1] "Abasic_1"  "Abasic_10" "Abasic_11" "Abasic_12" "Abasic_13"

# subset with `[` using `[features, cell_ids, experiments]`
features <- c("Uracil_45", "TP53")
cell_ids <- c("TGCGGGTGTAGAGTGC", "CTACGTCCACCACGTG")

fsce_small[features, cell_ids, ]

#> A FunctionalSingleCellExperiment object of 2 listed
#>  experiments with user-defined names and respective classes. 
#>  Containing an ExperimentList class object of length 2: 
#>  [1] rnaseq: SingleCellExperiment with 1 rows and 2 columns 
#>  [2] haircut: SingleCellExperiment with 1 rows and 2 columns 
#> Features: 
#>  experiments() - obtain the ExperimentList instance 
#>  colData() - the primary/phenotype DataFrame 
#>  sampleMap() - the sample availability DataFrame 
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment 
#>  *Format() - convert into a long or wide DataFrame 
#>  assays() - convert ExperimentList to a SimpleList of matrices

# dimensionality reduction results
SingleCellExperiment::reducedDimNames(fsce_small[["rnaseq"]])

#> [1] "PCA"  "UMAP" "TSNE"

# UMAP results (first 5 rows)
SingleCellExperiment::reducedDim(fsce_small[["rnaseq"]], "UMAP")[1:5, ]

#>                       [,1]       [,2]
#> TGCGGGTGTAGAGTGC  3.501003   3.623493
#> TGGTTCCCATCTATGG -2.851359   3.577775
#> CATATTCCACGTCAGC  3.724043   4.991723
#> TCACAAGTCCTGCTTG -6.166327 -12.030005
#> GGAATAATCCAGGGCT -4.296164   3.478746

# k-means cluster IDs
SingleCellExperiment::colData(fsce_small[["rnaseq"]])

#> DataFrame with 250 rows and 4 columns
#>                           cell_id   k_cluster leiden_cluster  cell_cycle
#>                       <character> <character>    <character> <character>
#> TGCGGGTGTAGAGTGC TGCGGGTGTAGAGTGC           5              2           S
#> TGGTTCCCATCTATGG TGGTTCCCATCTATGG           1              4           S
#> CATATTCCACGTCAGC CATATTCCACGTCAGC           2              2         G2M
#> TCACAAGTCCTGCTTG TCACAAGTCCTGCTTG           3              3          G1
#> GGAATAATCCAGGGCT GGAATAATCCAGGGCT           1              4          G1
#> ...                           ...         ...            ...         ...
#> CTACGTCCACCACGTG CTACGTCCACCACGTG           1              4          G1
#> TGTGGTAAGGTGCTTT TGTGGTAAGGTGCTTT           6              6          G1
#> AGCGTCGGTCGAGTTT AGCGTCGGTCGAGTTT           6              6          G1
#> CTCGAAAAGTTCGATC CTCGAAAAGTTCGATC           1              5         G2M
#> TGACAACGTCGCATAT TGACAACGTCGCATAT           1              4           S