Plot cells in reduced dimensionality 2D space

Cells can be colored by gene or feature in meta.data dataframe

plot_feature(fce, feature = NULL, expt = "sce", method = "UMAP",
  plot_dat = NULL, pt_size = 0.01, pt_alpha = 1,
  label_text = FALSE, label_size = 6, label_color = "grey",
  .cols = NULL, cell_filter = NULL, palette_type = "cloupe",
  col_pal = "Reds", max_y = NULL, plot_title = NULL,
  dims_to_plot = c(1, 2))

Arguments

fce
fce object of class MultiAssayExperiment

feature
single feature to plot, either gene name or column in colData

expt
Data type to use for plot, one of either sce (rna data, the default) or fsce (functional data). Defaults to "sce"

method
dimensionality reduction for plotting. defaults to UMAP

plot_dat
supplemental data.frame containing feature to plot. Must have a column named cell that contains matching colnames in colData

pt_size
size of points produced by geom_point

pt_alpha
alpha value for points plotted by geom_point

label_text
if TRUE display feature labels on plot

label_size
size of label text

label_color
color of label text

.cols
vector of colors to use for plot.

cell_filter
character vector of cell names to include in plot

palette_type
color palette type to use (either viridis, brewer, or cloupe) defaults to using cellranger loupe-like colors

col_pal
palette name to use if palette_type is brewer

max_y
maximum feature value to set scale to. Defaults to max of the feature

plot_title
string to supply for title for the plot, otherwise title is the feature

dims_to_plot
number of dimensions