Tidy reducedDims data

Tidy reducedDims data

tidy_dims(sce, dimnames = NULL, dims = c(1, 2))

Arguments

sce
An object of class SingleCellExperiment::SingleCellExperiment.

dimnames
vector of dimnames to retrieve. If NULL, retrieve all dimnames.

dims
vector of dimensions to retrieve

Tidying

This is a data tidier for a SingleCellExperient. Returns data from an sce in a tidy format, where variables are columns and observations are rows.

See also

Examples

tidy_dims(fsce_small)

#> # A tibble: 250 x 8
#>    experiment cell_id             PCA1   PCA2  UMAP1  UMAP2    TSNE1 TSNE2
#>    <chr>      <chr>              <dbl>  <dbl>  <dbl>  <dbl>    <dbl> <dbl>
#>  1 rnaseq     TGCGGGTTCAACTCTT   1.50   -9.31  2.15   3.99    6.93   -7.39
#>  2 rnaseq     TGGTTCCTCCCAAGAT   1.56   -6.91  1.38   3.64    5.02   -7.94
#>  3 rnaseq     CATATTCCATGCCCGA   2.26   -6.41  3.82   2.70   10.9    -3.63
#>  4 rnaseq     TCACAAGTCTCCAACC  -1.45   -1.06 -0.216  0.207  -0.0268  2.79
#>  5 rnaseq     GGACAAGAGCCCTAAT   3.57   -5.79  2.57   3.06    8.21   -1.86
#>  6 rnaseq     CTTTGCGAGCAGGTCA   1.61   -5.42  3.53   2.52   10.5    -2.55
#>  7 rnaseq     GTGCAGCAGTAGCCGA   3.25   -8.77  1.91   5.17    5.02   -5.49
#>  8 rnaseq     AGCGGTCCAATAGAGT  22.3    11.8   2.45  -1.47    8.55    6.08
#>  9 rnaseq     GGATGTTTCTCAAACG   0.953  -8.13  1.62   3.88    5.59   -8.33
#> 10 rnaseq     GTACTCCTCGTTTAGG  -3.35    1.37 -1.79  -1.51   -3.29    5.50
#> # … with 240 more rows

tidy_dims(fsce_small, dimnames = c("UMAP"))

#> # A tibble: 250 x 4
#>    experiment cell_id           UMAP1  UMAP2
#>    <chr>      <chr>             <dbl>  <dbl>
#>  1 rnaseq     TGCGGGTTCAACTCTT  2.15   3.99 
#>  2 rnaseq     TGGTTCCTCCCAAGAT  1.38   3.64 
#>  3 rnaseq     CATATTCCATGCCCGA  3.82   2.70 
#>  4 rnaseq     TCACAAGTCTCCAACC -0.216  0.207
#>  5 rnaseq     GGACAAGAGCCCTAAT  2.57   3.06 
#>  6 rnaseq     CTTTGCGAGCAGGTCA  3.53   2.52 
#>  7 rnaseq     GTGCAGCAGTAGCCGA  1.91   5.17 
#>  8 rnaseq     AGCGGTCCAATAGAGT  2.45  -1.47 
#>  9 rnaseq     GGATGTTTCTCAAACG  1.62   3.88 
#> 10 rnaseq     GTACTCCTCGTTTAGG -1.79  -1.51 
#> # … with 240 more rows