An example FunctionalSingleCellExperiment

Contains sample data for a functional single-cell experiment.

fsce_small

Format

An object of class FunctionalSingleCellExperiment of length 2.

Source

Generated by data-raw/fsce_small.R.

Details

There are two SingleCellExperiments:

rnaseq: a SingleCellExperiment object containing a single-cell mRNA sequencing (10x Genomics V2 3 prime) cell/gene counts matrix in the counts slot. Log-normalized counts are in the logcounts slot.

haircut: a SingleCellExperiment object containing a Haircut experiment using DNA repair oligos. Counts are in the counts slot, and centered log-ratio normalized counts are in the logcounts slot.

Haircut repair substrates include:

A "normal" DNA oligo (Normal_)

a ribonucleotide (rG) at position 45 (Ribo_)

an Abasic site at position 45 (Abasic_)

a uracil:adenosine base-pair, with uracil at position 45 (Uracil_)

a uracil:guanosine base-pair, with uracil at position 45 (GU_)

Dimensionality reduction calculations are in the rnaseq experiment, including PCA, UMAP, and TSNE slots. In addition, k-means clusters are available as rowData.

Examples

fsce_small

#> A FunctionalSingleCellExperiment object of 2 listed
#>  experiments with user-defined names and respective classes. 
#>  Containing an ExperimentList class object of length 2: 
#>  [1] rnaseq: SingleCellExperiment with 9479 rows and 250 columns 
#>  [2] haircut: SingleCellExperiment with 426 rows and 250 columns 
#> Features: 
#>  experiments() - obtain the ExperimentList instance 
#>  colData() - the primary/phenotype DataFrame 
#>  sampleMap() - the sample availability DataFrame 
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment 
#>  *Format() - convert into a long or wide DataFrame 
#>  assays() - convert ExperimentList to a SimpleList of matrices

# Individual experiments
fsce_small[["rnaseq"]]

#> class: SingleCellExperiment 
#> dim: 9479 250 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames(9479): FO538757.2 AP006222.2 ... AC004556.1 AC240274.1
#> rowData names(0):
#> colnames(250): TGCGGGTTCAACTCTT TGGTTCCTCCCAAGAT ... CTCGAAACACACTGCG
#>   TGACGGCAGAGTCGGT
#> colData names(2): cell_id k_cluster
#> reducedDimNames(3): PCA UMAP TSNE
#> spikeNames(0):

fsce_small[["haircut"]]

#> class: SingleCellExperiment 
#> dim: 426 250 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames(426): Abasic_1 Abasic_10 ... riboG_8 riboG_9
#> rowData names(2): hairpin position
#> colnames(250): TGCGGGTTCAACTCTT TGGTTCCTCCCAAGAT ... CTCGAAACACACTGCG
#>   TGACGGCAGAGTCGGT
#> colData names(1): cell_id
#> reducedDimNames(0):
#> spikeNames(0):

# Gene and activity names (first 5 items)
rownames(fsce_small[["rnaseq"]])[1:5]

#> [1] "FO538757.2" "AP006222.2" "NOC2L"      "ISG15"      "TNFRSF18"  

rownames(fsce_small[["haircut"]])[1:5]

#> [1] "Abasic_1"  "Abasic_10" "Abasic_11" "Abasic_12" "Abasic_13"

# subset with `[` using `[features, cell_ids, experiments]`
features <- c("Uracil_45", "TP53")
cell_ids <- c("GAAACTCTCCGAAGAG", "CGCTATCGTTCACCTC")

fsce_small[features, cell_ids, ]

#> Warning: Not all identifiers found in data

#> A FunctionalSingleCellExperiment object of 2 listed
#>  experiments with user-defined names and respective classes. 
#>  Containing an ExperimentList class object of length 2: 
#>  [1] rnaseq: SingleCellExperiment with 1 rows and 1 columns 
#>  [2] haircut: SingleCellExperiment with 1 rows and 1 columns 
#> Features: 
#>  experiments() - obtain the ExperimentList instance 
#>  colData() - the primary/phenotype DataFrame 
#>  sampleMap() - the sample availability DataFrame 
#>  `$`, `[`, `[[` - extract colData columns, subset, or experiment 
#>  *Format() - convert into a long or wide DataFrame 
#>  assays() - convert ExperimentList to a SimpleList of matrices

library(SingleCellExperiment)

#> Loading required package: SummarizedExperiment

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following object is masked from ‘package:scrunchy’:
#> 
#>     normalize

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colMeans,
#>     colnames, colSums, dirname, do.call, duplicated, eval, evalq,
#>     Filter, Find, get, grep, grepl, intersect, is.unsorted, lapply,
#>     lengths, Map, mapply, match, mget, order, paste, pmax, pmax.int,
#>     pmin, pmin.int, Position, rank, rbind, Reduce, rowMeans, rownames,
#>     rowSums, sapply, setdiff, sort, table, tapply, union, unique,
#>     unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> Loading required package: BiocParallel

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from ‘package:base’:
#> 
#>     aperm, apply

# dimensionality reduction results
reducedDimNames(fsce_small[["rnaseq"]])

#> [1] "PCA"  "UMAP" "TSNE"

# UMAP results (first 5 rows)
reducedDim(fsce_small[["rnaseq"]], "UMAP")[1:5, ]

#>                       [,1]      [,2]
#> TGCGGGTTCAACTCTT  2.146147 3.9898103
#> TGGTTCCTCCCAAGAT  1.376538 3.6447333
#> CATATTCCATGCCCGA  3.815146 2.7023185
#> TCACAAGTCTCCAACC -0.216390 0.2069193
#> GGACAAGAGCCCTAAT  2.566830 3.0588671

# k-means cluster IDs
colData(fsce_small[["rnaseq"]])

#> DataFrame with 250 rows and 2 columns
#>                           cell_id   k_cluster
#>                       <character> <character>
#> TGCGGGTTCAACTCTT TGCGGGTTCAACTCTT           6
#> TGGTTCCTCCCAAGAT TGGTTCCTCCCAAGAT           6
#> CATATTCCATGCCCGA CATATTCCATGCCCGA           6
#> TCACAAGTCTCCAACC TCACAAGTCTCCAACC           6
#> GGACAAGAGCCCTAAT GGACAAGAGCCCTAAT           6
#> ...                           ...         ...
#> CTACGTCCAGTGGAGT CTACGTCCAGTGGAGT           4
#> TTAGGCACAGGTGGAT TTAGGCACAGGTGGAT           2
#> AGCGTCGTCGTCGTTC AGCGTCGTCGTCGTTC           4
#> CTCGAAACACACTGCG CTCGAAACACACTGCG           6
#> TGACGGCAGAGTCGGT TGACGGCAGAGTCGGT           6